Add minimal working version of the Open With addin
Adds a Preferences... menu to the Open With context menu which
opens a dialog that lists the available editors and applications.
The dialog can be used to change the default application or editor.

IExternalDisplayBindings are not currently supported, but
applications and other display bindings are supported, so you can,
for example change a text file so it opens in the Hex Editor by
default, or opens with the TextEdit.app. Once configured the
new editor or application will be shown as the first item in the
right click context menu.

Configuration is currently not saved when the IDE is restarted.
The Open With dialog's Remove and Add button are not implemented.