Fix duplicate items in Open With dialog
When an editor or application has been configured to be the default
when the Open With dialog was opened again the editor or application
would be duplicated in the list. To fix this the custom runtime
display bindings that have been added are ignored when creating
the list in the Open With dialog.