Remove some code that is not required
External display bindings seem to work already. They use a
DesktopApplication so the existing code handles this. Removed the
custom external display binding class.

The only external display binding I can find is the one that
opens storyboards or .xib files into Xcode. This display binding
makes extra checks to see if it supports the file as well as
checking the filename so just relying on calling the Xcode
application may mean it can be opened when Visual Studio for Mac
would disable it based on these extra checks. For now this problem
is going to be ignored.