Save configured Open With applications
The current settings are saved to an xml file under the user's
profile. This file is read on startup and the Open With settings are
re-applied.

Currently changing the default Open With to map to an existing
desktop application that is returned by the IDE does not work.
Custom applications and view display bindings that are made the
default are working.