Default applications now configured on re-opening the IDE
Making a file open with an external application, such as TextEdit.app,
would not be remembered on re-opening the IDE. The problem was that
the full path to an existing filename is needed in order to get
the associated desktop application as a display binding. To handle
this there is now a lazy display binding created for these mapped
applications. When a file is opened in the IDE then the associated
desktop application will be found at that point instead of on IDE
startup.